refactor(sign): Add isValid functions in functions.js
